requirements.txt
langchain
langchain-community
langchain-openai
streamlit
python-dotenv